Rename package to communities
