+++ communities/src/androidTest/java/com/neaniesoft/vermilion/communities/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.neaniesoft.vermilion.communitylist
+package com.neaniesoft.vermilion.communities
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
